Release terminal runner reservations after final leases
