+++ libs/api/runners/src/db/job-executions.ts
@@ -39,6 +39,7 @@ import {
 } from '#metrics/instance.js';
 import type {Tx} from './db.js';
 import {db} from './db.js';
+import {releaseTerminalRunnerInstanceReservationsByIds} from './reservations.js';
 import {runnersOutbox} from './schema/outbox.js';
 import {pendingJobExecutions} from './schema/pending-job-executions.js';
 import {providerRunners} from './schema/runner-instances.js';
@@ -405,6 +406,8 @@ async function touchRunnerSessionLiveness(params: {
 /**
  * Releases a job execution's lease when the orchestration workflow finalizes it: deletes the
  * running-job-execution row AND any lingering pending row for the same execution, in one tx.
+ * When the deleted lease was the last one for a terminal provider runner, it also releases the
+ * runner's reservation in the same transaction.
  * Idempotent (0-row no-op), no token scope (the workflow is authoritative), and
  * emits no event: the workflow already owns the outcome. Sweeping the pending row
  * too closes the at-least-once window where an enqueue retry left an orphan that a
@@ -421,9 +424,29 @@ export async function releaseJobExecution(params: {jobExecutionId: string}): Pro
     await tx
       .delete(pendingJobExecutions)
       .where(eq(pendingJobExecutions.jobExecutionId, params.jobExecutionId));
-    await tx
+    const deletedRunningRows = await tx
       .delete(runningJobExecutions)
-      .where(eq(runningJobExecutions.jobExecutionId, params.jobExecutionId));
+      .where(eq(runningJobExecutions.jobExecutionId, params.jobExecutionId))
+      .returning({
+        workspaceId: runningJobExecutions.workspaceId,
+        provisionerId: runningJobExecutions.provisionerId,
+        providerRunnerId: runningJobExecutions.providerRunnerId,
+      });
+
+    const deletedRunningRow = deletedRunningRows[0];
+    if (deletedRunningRow?.provisionerId && deletedRunningRow.providerRunnerId) {
+      // Serialize this cleanup with terminal reports and the stale-runner reaper, which use the
+      // workspace lock while inspecting the runner and its leases.
+      await tx.execute(
+        sql`select pg_advisory_xact_lock(hashtext(${deletedRunningRow.workspaceId}))`,
+      );
+      await releaseTerminalRunnerInstanceReservationsByIds(tx, {
+        workspaceId: null,
+        provisionerId: deletedRunningRow.provisionerId,
+        providerRunnerIds: [deletedRunningRow.providerRunnerId],
+        requireUnlinkedSession: false,
+      });
+    }
   });
 }
