+++ libs/api/runners/src/db/runner-instances.ts
@@ -10,7 +10,6 @@ import {
   isNotNull,
   isNull,
   lt,
-  lte,
   notExists,
   notInArray,
   or,
@@ -29,7 +28,7 @@ import {
   type RunnerInstanceBoundJobExecution,
 } from './job-executions.js';
 import {lockRunnerReservationAdvisoryKeysTx} from './reservation-locks.js';
-import {releaseReservationUnits} from './reservations.js';
+import {releaseTerminalRunnerInstanceReservationsByIds} from './reservations.js';
 import {validateRunnerReservationCapacityTx} from './runner-assignments.js';
 import {activeStates, terminalStates} from './runner-states.js';
 import {ephemeralRegistrationTokens} from './schema/ephemeral-registration-tokens.js';
@@ -916,221 +915,6 @@ async function releaseTerminalRunnerInstanceReservations(
     ),
     requireUnlinkedSession: false,
   });
-}
-
-export async function releaseTerminalRunnerInstanceReservationsByIds(
-  tx: Tx,
-  params: {
-    workspaceId: string | null;
-    provisionerId: string;
-    providerRunnerIds?: string[];
-    runnerInstanceIds?: string[];
-    reportedAtByProviderRunnerId?: ReadonlyMap<string, Date>;
-    requireUnlinkedSession?: boolean;
-  },
-): Promise<number> {
-  if (
-    (params.providerRunnerIds?.length ?? 0) === 0 &&
-    (params.runnerInstanceIds?.length ?? 0) === 0
-  )
-    return 0;
-
-  const reservationWorkspacePredicate =
-    params.workspaceId === null
-      ? sql``
-      : sql`and ${eq(reservations.workspaceId, params.workspaceId)}`;
-  const noRunningJobPredicate = notExists(
-    tx
-      .select({id: runningJobExecutions.id})
-      .from(runningJobExecutions)
-      .where(
-        and(
-          eq(runningJobExecutions.provisionerId, params.provisionerId),
-          eq(runningJobExecutions.providerRunnerId, providerRunners.providerRunnerId),
-          params.workspaceId === null
-            ? undefined
-            : eq(runningJobExecutions.workspaceId, params.workspaceId),
-        ),
-      ),
-  );
-  const reportFreshnessPredicate =
-    params.reportedAtByProviderRunnerId && params.reportedAtByProviderRunnerId.size > 0
-      ? or(
-          ...[...params.reportedAtByProviderRunnerId].map(([providerRunnerId, reportedAt]) =>
-            and(
-              eq(providerRunners.providerRunnerId, providerRunnerId),
-              lte(providerRunners.reportedAt, reportedAt),
-            ),
-          ),
-        )
-      : undefined;
-  const runnerIdentityPredicate = or(
-    params.providerRunnerIds && params.providerRunnerIds.length > 0
-      ? inArray(providerRunners.providerRunnerId, params.providerRunnerIds)
-      : undefined,
-    params.runnerInstanceIds && params.runnerInstanceIds.length > 0
-      ? inArray(providerRunners.id, params.runnerInstanceIds)
-      : undefined,
-  );
-
-  const rows = await tx
-    .select({
-      id: providerRunners.id,
-      releaseReservationId: sql<string | null>`coalesce(
-        (select ${reservations.id}
-         from ${reservations}
-         where ${reservations.id} = ${providerRunners.intendedReservationId}
-           and ${reservations.provisionerId} = ${params.provisionerId}
-           ${reservationWorkspacePredicate}),
-        (select ${reservations.id}
-         from ${reservations}
-         where ${reservations.id} = ${providerRunners.reservationId}
-           and ${reservations.provisionerId} = ${params.provisionerId}
-           ${reservationWorkspacePredicate})
-      )`,
-      releaseReservationWorkspaceId: sql<string | null>`coalesce(
-        (select ${reservations.workspaceId}
-         from ${reservations}
-         where ${reservations.id} = ${providerRunners.intendedReservationId}
-           and ${reservations.provisionerId} = ${params.provisionerId}
-           ${reservationWorkspacePredicate}),
-        (select ${reservations.workspaceId}
-         from ${reservations}
-         where ${reservations.id} = ${providerRunners.reservationId}
-           and ${reservations.provisionerId} = ${params.provisionerId}
-           ${reservationWorkspacePredicate})
-      )`,
-    })
-    .from(providerRunners)
-    .where(
-      and(
-        eq(providerRunners.provisionerId, params.provisionerId),
-        runnerIdentityPredicate,
-        inArray(providerRunners.state, terminalStates),
-        or(
-          isNotNull(providerRunners.reservationId),
-          isNotNull(providerRunners.intendedReservationId),
-        ),
-        params.workspaceId === null
-          ? undefined
-          : or(
-              and(
-                eq(providerRunners.workspaceId, params.workspaceId),
-                or(
-                  exists(
-                    tx
-                      .select({id: reservations.id})
-                      .from(reservations)
-                      .where(
-                        and(
-                          eq(reservations.workspaceId, params.workspaceId),
-                          eq(reservations.provisionerId, params.provisionerId),
-                          eq(reservations.id, providerRunners.reservationId),
-                        ),
-                      ),
-                  ),
-                  exists(
-                    tx
-                      .select({id: reservations.id})
-                      .from(reservations)
-                      .where(
-                        and(
-                          eq(reservations.workspaceId, params.workspaceId),
-                          eq(reservations.provisionerId, params.provisionerId),
-                          eq(reservations.id, providerRunners.intendedReservationId),
-                        ),
-                      ),
-                  ),
-                ),
-              ),
-              and(
-                isNull(providerRunners.workspaceId),
-                isNotNull(providerRunners.intendedReservationId),
-                exists(
-                  tx
-                    .select({id: reservations.id})
-                    .from(reservations)
-                    .where(
-                      and(
-                        eq(reservations.workspaceId, params.workspaceId),
-                        eq(reservations.provisionerId, params.provisionerId),
-                        eq(reservations.id, providerRunners.intendedReservationId),
-                      ),
-                    ),
-                ),
-              ),
-            ),
-        params.requireUnlinkedSession === false
-          ? undefined
-          : isNull(providerRunners.runnerSessionId),
-        noRunningJobPredicate,
-        reportFreshnessPredicate,
-        isNull(providerRunners.reservationReleasedAt),
-      ),
-    )
-    .for('update');
-
-  if (rows.length === 0) return 0;
-
-  const updated = await tx
-    .update(providerRunners)
-    .set({
-      intendedReservationId: null,
-      reservationReleasedAt: sql`now()`,
-      updatedAt: sql`now()`,
-    })
-    .where(
-      and(
-        inArray(
-          providerRunners.id,
-          rows.map((row) => row.id),
-        ),
-        params.requireUnlinkedSession === false
-          ? undefined
-          : isNull(providerRunners.runnerSessionId),
-        noRunningJobPredicate,
-        reportFreshnessPredicate,
-        isNull(providerRunners.reservationReleasedAt),
-      ),
-    )
-    .returning({id: providerRunners.id});
-
-  const releasesByReservationId = new Map<
-    string,
-    {workspaceId: string; reservationId: string; count: number}
-  >();
-  const updatedIds = new Set(updated.map((row) => row.id));
-  for (const row of rows) {
-    if (!updatedIds.has(row.id)) continue;
-    if (!row.releaseReservationId || !row.releaseReservationWorkspaceId) continue;
-    const key = `${row.releaseReservationWorkspaceId}:${row.releaseReservationId}`;
-    const release = releasesByReservationId.get(key) ?? {
-      workspaceId: row.releaseReservationWorkspaceId,
-      reservationId: row.releaseReservationId,
-      count: 0,
-    };
-    release.count += 1;
-    releasesByReservationId.set(key, release);
-  }
-
-  if (releasesByReservationId.size === 0) return 0;
-
-  const releasesByWorkspaceId = new Map<string, Array<{reservationId: string; count: number}>>();
-  for (const release of releasesByReservationId.values()) {
-    const releases = releasesByWorkspaceId.get(release.workspaceId) ?? [];
-    releases.push({reservationId: release.reservationId, count: release.count});
-    releasesByWorkspaceId.set(release.workspaceId, releases);
-  }
-
-  let released = 0;
-  for (const [workspaceId, releases] of releasesByWorkspaceId) {
-    released += await releaseReservationUnits(tx, {
-      workspaceId,
-      provisionerId: params.provisionerId,
-      releases,
-    });
-  }
-  return released;
 }
 
 function staleRunnerInstanceCutoff(thresholdSeconds: number): SQL {
